move cancelableResource.ts into a browser subfolder to avoid mixing with Puppeteer tests
