+++ packages/calcite-components/src/components/action-bar/action-bar.browser.e2e.tsx
@@ -1,6 +1,6 @@
 import { describe } from "vitest";
 import { mockConsole } from "../../tests/utils/logging";
-import { cancelable } from "../../tests/commonTests/cancelableResource";
+import { cancelable } from "../../tests/commonTests/browser/cancelableResource";
 
 describe("action-bar", () => {
   mockConsole("warn");
