+++ packages/calcite-components/src/components/block-group/block-group.browser.e2e.tsx
@@ -1,6 +1,6 @@
 import { describe } from "vitest";
 import { mockConsole } from "../../tests/utils/logging";
-import { cancelable } from "../../tests/commonTests/cancelableResource";
+import { cancelable } from "../../tests/commonTests/browser/cancelableResource";
 
 describe("block-group", () => {
   mockConsole("warn");
