+++ packages/calcite-components/src/components/color-picker/color-picker.browser.e2e.ts
@@ -1,6 +1,6 @@
 import { describe } from "vitest";
 import { mockConsole } from "../../tests/utils/logging";
-import { cancelable } from "../../tests/commonTests/cancelableResource";
+import { cancelable } from "../../tests/commonTests/browser/cancelableResource";
 
 describe("color-picker", () => {
   mockConsole("warn");
